refactor(4770-s2): fix layering-inversion runtime imports via TYPE_CHECKING
Fix upward/runtime import dependencies flagged in the #4770 survey:

1. simulator.py:38-39 — gym_env config imported at RUNTIME (upward dep).
   Converted to TYPE_CHECKING-guarded import mirroring sim/facade.py:16
   pattern: added  and moved
   EnvSettings, PedEnvSettings, SimulationSettings, PedestrianSimulationConfig,
   RobotSimulationConfig behind .

2. baseline_probabilistic_predictor.py:20 — nav imports benchmark at module
   level. Restructured so nav doesn't import benchmark:
   - Moved PedestrianState and NeighborContext to nav/predictive_types.py
     (canonical nav-level pedestrian state types).
   - benchmark/pedestrian_forecast.py re-exports both for backward compat.
   - ForecastBaselineFunction type alias moved behind TYPE_CHECKING.
   - Baseline function imports made lazy inside _baseline_for_variant().

Behavior-preservation:
- Full test suite green (759 passed, 3 pre-existing collection errors from
  missing optional deps duckdb/pyarrow).
- Runtime import + instantiation of Simulator, BaselineProbabilisticPredictor,
  PedestrianState, NeighborContext confirmed — no NameError or import cycle.
- Zero functional test changes needed; only monkeypatch targets updated in
  test_baseline_probabilistic_predictor.py to match new lazy-import structure.

Refs #4770
